quick commit

+++ app/Http/Controllers/CustomCampaignsController.php
@@ -94,33 +94,35 @@ class CustomCampaignsController extends Controller
 
 
                     $user =  CampaignUser::where('campaign_system_id', $systemNode->id)->first();
-                    $user->update(['campaign_system_id' => null, 'status_id' => null]);
-                    $message = CampaignUserRecords::where('id', $user->id)->first();
-                    $flag = collect([
-                        'message' => $message,
-                        'id' => $campid
-                    ]);
-                    broadcast(new CampaignUserUpdate($flag));
+                    if ($user->count() > 0) {
+                        $user->update(['campaign_system_id' => null, 'status_id' => null]);
+                        $message = CampaignUserRecords::where('id', $user->id)->first();
+                        $flag = collect([
+                            'message' => $message,
+                            'id' => $campid
+                        ]);
+                        broadcast(new CampaignUserUpdate($flag));
 
-                    if ($systemNode->node_join_count > 0) {
-                        $nodes = NodeJoin::where('campaign_system_id', $systemNode->id);
-                        foreach ($nodes as $node) {
-                            $flag = null;
-                            $flag = collect([
-                                'joinNodeID' => $node->id,
-                                'id' => $campid
-                            ]);
-                            broadcast(new NodeJoinDelete($flag));
+                        if ($systemNode->node_join_count > 0) {
+                            $nodes = NodeJoin::where('campaign_system_id', $systemNode->id);
+                            foreach ($nodes as $node) {
+                                $flag = null;
+                                $flag = collect([
+                                    'joinNodeID' => $node->id,
+                                    'id' => $campid
+                                ]);
+                                broadcast(new NodeJoinDelete($flag));
 
-                            $user = CampaignUser::where('campaign_system_id', $node->campaign_system_id)->first();
-                            $user->update(['campaign_system_id' => null, 'status_id' => null]);
-                            $message = CampaignUserRecords::where('id', $user->id)->first();
-                            $flag = collect([
-                                'message' => $message,
-                                'id' => $campid
-                            ]);
-                            broadcast(new CampaignUserUpdate($flag));
-                            $node->delete();
+                                $user = CampaignUser::where('campaign_system_id', $node->campaign_system_id)->first();
+                                $user->update(['campaign_system_id' => null, 'status_id' => null]);
+                                $message = CampaignUserRecords::where('id', $user->id)->first();
+                                $flag = collect([
+                                    'message' => $message,
+                                    'id' => $campid
+                                ]);
+                                broadcast(new CampaignUserUpdate($flag));
+                                $node->delete();
+                            }
                         }
                     }
